unify all validators in App
