+++ src/components/views/verification/projectContact/common.tsx
@@ -2,9 +2,9 @@ import { IconLink, neutralColors } from '@giveth/ui-design-system';
 import { Dispatch, FC, SetStateAction } from 'react';
 import styled from 'styled-components';
 import Input, { IFormValidations } from '@/components/Input';
-import { regexList } from '@/lib/helpers';
 import { FlexCenter } from '@/components/styled-components/Flex';
 import { RemoveIcon } from '@/components/views/verification/common';
+import { validators } from '@/lib/constants/regex';
 
 interface IProps {
 	setUrl?: (i: string) => void;
@@ -32,14 +32,7 @@ export const WebsiteInput: FC<IProps> = ({
 				LeftIcon={<IconLink color={neutralColors.gray[600]} />}
 				disabled={!setUrl}
 				setFormValidation={setFormValidation}
-				validators={
-					setUrl && [
-						{
-							pattern: regexList.url,
-							msg: 'Invalid URL',
-						},
-					]
-				}
+				validators={setUrl && [validators.url]}
 			/>
 			{!setUrl && <RemoveIcon onClick={() => remove && remove()} />}
 		</Container>
